+++ apps/api/src/controllers/v2/types.ts
@@ -17,35 +17,7 @@ import {
 import type { InternalOptions } from "../../scraper/scrapeURL";
 import { ErrorCodes } from "../../lib/error";
 import Ajv from "ajv";
-
-export enum IntegrationEnum {
-  DIFY = "dify",
-  ZAPIER = "zapier",
-  PIPEDREAM = "pipedream",
-  RAYCAST = "raycast",
-  LANGCHAIN = "langchain",
-  CREWAI = "crewai",
-  LLAMAINDEX = "llamaindex",
-  N8N = "n8n",
-  CAMELAI = "camelai",
-  MAKE = "make",
-  FLOWISE = "flowise",
-  METAGPT = "metagpt",
-  RELEVANCEAI = "relevanceai",
-}
-
-export const integrationSchema = z
-  .string()
-  .refine(
-    val =>
-      (typeof val === "string" && val.startsWith("_")) ||
-      Object.values(IntegrationEnum).includes(val as any),
-    {
-      message: `Invalid enum value. Expected ${Object.values(IntegrationEnum)
-        .map(v => `'${v}'`)
-        .join(" | ")}`,
-    },
-  );
+import { integrationSchema } from "../../utils/integration";
 
 export type Format =
   | "markdown"
